requirements.txt
flask
flask_restful
flask-cors
pylint
psycopg2
Flask-SQLAlchemy
sqlalchemy-utils
marshmallow